# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Signara API
  version: 1.0.0
  description: |
    Secure Every Signature. — REST API for the Signara digital document signing and
    agreement management platform.

    - Base path: `/api/v1`
    - Interactive docs (Swagger UI): `/api/v1/docs`
    - Live YAML: `/api/v1/openapi.yaml`

    ## Authentication
    Two mechanisms are supported:

    | Scheme | Where | Usage |
    | --- | --- | --- |
    | Bearer JWT (`access-token`) | `Authorization: Bearer <token>` | Interactive sessions issued by the OIDC flow |
    | API key (`api-key`) | `X-API-Key: sgn_...` | Machine-to-machine clients |

    ## Tenancy
    Tenant-scoped endpoints require an active organization context. Endpoints decorated
    with `x-tenant-required: true` return `403` when the caller has no membership.

    ## Idempotency
    Mutating endpoints accept `X-Idempotency-Key` — repeat requests with the same key
    return the stored response instead of re-executing.
  termsOfService: https://signara.innotel.us/terms
  contact:
    name: Signara
    url: https://signara.innotel.us
    email: support@signara.innotel.us
  license:
    name: AGPL-3.0-or-later
    identifier: AGPL-3.0-or-later
servers:
  - url: https://api.signara.innotel.us/api/v1
    description: Production
  - url: http://localhost:8000/api/v1
    description: Local development
...
security:
  - access-token: []
paths:
...
  /documents:
    get:
      tags: [documents]
      x-tenant-required: true
      summary: List documents (searches via Meilisearch when `search` is set)
      parameters:
        - { name: search, in: query, schema: { type: string } }
        - { name: status, in: query, schema: { $ref: '#/components/schemas/DocumentStatus' } }
        - { name: workspaceId, in: query, schema: { type: string, format: uuid } }
        - { name: limit, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Paginated documents
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentList' }
  /documents/upload:
    post:
      tags: [documents]
      x-tenant-required: true
      summary: Upload a document (PDF, DOCX, PNG, JPEG, WebP; max 50 MB)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                title: { type: string }
                description: { type: string }
                workspaceId: { type: string, format: uuid }
                tags:
                  type: array
                  items: { type: string }
      responses:
        '201':
          description: Document created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
        '400':
          description: Unsupported type or payload too large
          $ref: '#/components/responses/BadRequest'
  /documents/{id}:
    get:
      tags: [documents]
      x-tenant-required: true
      summary: Get a document with version history
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Document detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentDetail' }
        '404': { $ref: '#/components/responses/NotFound' }
    patch:
      tags: [documents]
      x-tenant-required: true
      summary: Update document metadata
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title: { type: string }
                description: { type: string }
                tags:
                  type: array
                  items: { type: string }
                workspaceId: { type: string, format: uuid }
      responses:
        '200':
          description: Updated document
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Document' }
    delete:
      tags: [documents]
      x-tenant-required: true
      summary: Soft-delete a document
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Document deleted
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Success' }
  /documents/{id}/download:
    get:
      tags: [documents]
      x-tenant-required: true
      summary: Get a time-limited presigned download URL
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
        - { name: version, in: query, schema: { type: integer, minimum: 1 }, description: Optional version number }
      responses:
        '200':
          description: Presigned URL (valid 15 min)
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string, format: uri }
                  fileName: { type: string }
                  contentType: { type: string }
  /documents/{id}/versions:
    post:
      tags: [documents]
      x-tenant-required: true
      summary: Upload a new immutable version of a document
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
                changeNote: { type: string }
      responses:
        '201':
          description: New version recorded
          content:
            application/json:
              schema:
                type: object
                properties:
                  version: { $ref: '#/components/schemas/DocumentVersion' }
                  document: { $ref: '#/components/schemas/Document' }

  # ------------------------------------------------------------ templates ---
...
  /certificates:
    get:
      tags: [certificates]
      x-tenant-required: true
      summary: List certificates for the organization
      parameters:
        - { name: status, in: query, schema: { $ref: '#/components/schemas/CertificateStatus' } }
        - { name: limit, in: query, schema: { type: integer, default: 25, maximum: 100 } }
        - { name: offset, in: query, schema: { type: integer, default: 0 } }
      responses:
        '200':
          description: Paginated certificates
          content:
            application/json:
              schema:
                type: object
                properties:
                  total: { type: integer }
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/SigningCertificate' }
...
  /certificates/import:
    post:
      tags: [certificates]
      x-tenant-required: true
      summary: Import an existing certificate (enterprise PKI / internal CA)
      description: Optional private key is verified against the certificate and encrypted at rest with CRYPTO_MASTER_KEY. Re-importing a serial reactivates the row.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [certificatePem]
              properties:
                certificatePem: { type: string, description: End-entity certificate in PEM }
                chainPem: { type: string, description: Intermediates + root (PEM), optional }
                privateKeyPem: { type: string, description: Signing private key (PEM), optional — encrypted at rest }
                provider: { $ref: '#/components/schemas/CertificateProvider' }
                validationLevel: { type: string, enum: [DV, OV, EV] }
                commonName: { type: string }
      responses:
        '201':
          description: Imported certificate (private key never returned)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningCertificate' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /certificates/provision:
    post:
      tags: [certificates]
      x-tenant-required: true
      summary: Provision a certificate via ACME, Cerulean, or an internal PKI script
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, commonName]
              properties:
                provider: { $ref: '#/components/schemas/CertificateProvider' }
                commonName: { type: string, description: Subject CN — signer email for person certificates }
                email: { type: string, format: email }
                options: { type: object, description: Provider-specific (ACME dns handler, Cerulean template, ...) }
      responses:
        '201':
          description: Provisioned and stored certificate
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningCertificate' }
        '400': { $ref: '#/components/responses/BadRequest' }
  /certificates/verify:
    post:
      tags: [certificates]
      x-tenant-required: true
      summary: Verify a certificate and (optionally) a cryptographic signature
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                certificatePem: { type: string }
                certificateId: { type: string, format: uuid }
                signatureValue: { type: string, description: Base64 signature over the digest }
                signedHash: { type: string, description: SHA-256 hex of the signed content }
      responses:
        '200':
          description: Verification report with identity assurance
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CertificateVerificationResult' }
  /certificates/{id}:
    get:
      tags: [certificates]
      x-tenant-required: true
      summary: Get a certificate
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200':
          description: Certificate detail
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningCertificate' }
        '404': { $ref: '#/components/responses/NotFound' }
  /certificates/{id}/revoke:
    patch:
      tags: [certificates]
      x-tenant-required: true
      summary: Revoke a certificate (immediately blocks certificate-backed signing)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason: { type: string }
      responses:
        '200':
          description: Certificate revoked
          content:
            application/json:
              schema: { $ref: '#/components/schemas/SigningCertificate' }

  # ---------------------------------------------------------------- admin ---
...
components:
  securitySchemes:
    access-token:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT issued by the Signara API after the Authentik OIDC flow, or the raw Authentik access token.
    api-key:
      type: apiKey
      in: header
      name: X-API-Key
      description: "API key created via POST /api-keys, e.g. `X-API-Key: sgn_...`"
  responses:
    BadRequest:
      description: Validation or business rule violation
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
    NotFound:
      description: Resource does not exist or is outside the tenant
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorEnvelope' }
...
  schemas:
    Success:
      type: object
      properties:
        success: { type: boolean, example: true }
    ErrorEnvelope:
      type: object
      properties:
        statusCode: { type: integer }
        error: { type: string }
        message:
          type: array
          items: { type: string }
        path: { type: string }
        timestamp: { type: string, format: date-time }
        requestId: { type: string }

    # ------------------------------------------------------------- auth ---
...
    DocumentStatus:
      type: string
      enum: [DRAFT, AWAITING_SIGNATURE, IN_PROGRESS, COMPLETED, VOIDED, CANCELLED, EXPIRED]
    Document:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        workspaceId: { type: string, format: uuid, nullable: true }
        templateId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        description: { type: string, nullable: true }
        status: { $ref: '#/components/schemas/DocumentStatus' }
        fileName: { type: string }
        contentHostType: { type: string, nullable: true }
        sizeBytes: { type: integer, nullable: true }
        checksumSha256: { type: string, nullable: true }
        version: { type: integer, default: 1 }
        tags:
          type: array
          items: { type: string }
        metadata: { type: object, nullable: true }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    DocumentList:
      type: object
      properties:
        total: { type: integer }
        limit: { type: integer }
        offset: { type: integer }
        items:
          type: array
          items: { $ref: '#/components/schemas/Document' }
    DocumentVersion:
      type: object
      properties:
        id: { type: string, format: uuid }
        documentId: { type: string, format: uuid }
        version: { type: integer }
        sizeBytes: { type: integer, nullable: true }
        checksumSha256: { type: string, nullable: true }
        changeNote: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
    DocumentDetail:
      allOf:
        - $ref: '#/components/schemas/Document'
        - type: object
          properties:
            versions:
              type: array
              items: { $ref: '#/components/schemas/DocumentVersion' }

    # ------------------------------------------------------------ templates -
...
    IdentityAssurance:
      type: object
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: Transparent factor-weighted score (see docs/Security.md)
        level:
          type: string
          enum: [NONE, BASIC, ADVANCED, HIGH]
        factors:
          type: object
          properties:
            certificateValid: { type: boolean }
            certIdentityMatch: { type: boolean }
            accountVerified: { type: boolean }
            mfaEnabled: { type: boolean }
            revocationChecked: { type: boolean }
            providerValidation: { type: string, nullable: true, enum: [DV, OV, EV] }
    SigningCertificate:
      type: object
      properties:
        id: { type: string, format: uuid }
        organizationId: { type: string, format: uuid }
        userId: { type: string, format: uuid, nullable: true }
        provider: { $ref: '#/components/schemas/CertificateProvider' }
        status: { $ref: '#/components/schemas/CertificateStatus' }
        commonName: { type: string }
        email: { type: string, format: email, nullable: true }
        serialNumber: { type: string }
        subjectDN: { type: string }
        issuerDN: { type: string, nullable: true }
        notBefore: { type: string, format: date-time }
        notAfter: { type: string, format: date-time }
        certificatePem: { type: string, description: Public certificate — safe to display }
        keyAlgorithm: { type: string, nullable: true }
        validationLevel: { type: string, nullable: true, enum: [DV, OV, EV] }
        sourceRef: { type: string, nullable: true }
        revokedAt: { type: string, format: date-time, nullable: true }
        revokedReason: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
      description: Private keys are never returned; they are encrypted at rest (AES-256-GCM under CRYPTO_MASTER_KEY) or held by the provider.
    CertificateProvider:
      type: string
      enum: [CERULEAN, ACME, INTERNAL_PKI, WEB]
    CertificateStatus:
      type: string
      enum: [ACTIVE, REVOKED, EXPIRED]
    CertificateVerificationResult:
      type: object
      properties:
        valid: { type: boolean }
        serialNumber: { type: string }
        subjectDN: { type: string }
        issuerDN: { type: string, nullable: true }
        notBefore: { type: string, format: date-time }
        notAfter: { type: string, format: date-time }
        email: { type: string, nullable: true }
        signatureVerified: { type: boolean, nullable: true }
        identityAssurance: { $ref: '#/components/schemas/IdentityAssurance' }
        reasons:
          type: array
          items: { type: string }
...